LinoWeb/requirements.txt
alembic==1.0.11
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
decorator==4.4.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
idna==2.5
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
networkx==2.3
numpy==1.17.0
Pillow==6.1.0
python-dateutil==2.8.0
python-editor==1.0.4
PyWavelets==1.0.3
requests2==2.16.0
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
urllib3==1.21.1
Werkzeug==0.15.5
